requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
mysqlclient==2.2.7
sqlparse==0.5.3
